interim UG update, 7/10/18
this update contains the following:
1.  additional fields in the GUI dialog for gender and ethnicity
2. variables to store gender and ethnicity
3. stim_map currently references the olderadultMale_C and youngeradultMale_C emojis.

I still need to edit the code with some conditional statements such that depending on the combo of gender and ethnicity entered into the GUI dialog, different emojis will be used.  But I haven't figured out exactly where to place that yet without getting a bunch of errors.