trustee practice, 7.17.18
This commit contains code to run a practice script for the trustee/investee role, as well as the associated design file (6 trials).

this is not final, as currently if a suvbject does not respond, there is no 'you have 3 seconds to respond message' that appears.  but that will be added soon.